requirements.txt
streamlit
yfinance
pandas
numpy
scikit-learn
plotly
fpdf2
feedparser